Use PyCURL rather than wget to manage downloads
This results in a more complicated, but it adds several very useful features,
such as a "real-time" progress bar, which makes it possible to do new thinks
like computing the current download speed and the estimated time until a
download end.

Another possibility would have been to parse wget's output, but this is
difficult and unreliable. This solution is much safer and easier.